2024/1/22AEC修正分 ・休損型数 小数入力対応 データ型：int → decimal(13, 3)　※DB定義変更あり

+++ TMQ6WebSolution/BackEndSolution/BusinessLogic_MA0001/BusinessLogicDataClass_MA0001.cs
@@ -1575,7 +1575,7 @@ namespace BusinessLogic_MA0001
             public int? LossAbsence { get; set; }
             /// <summary>Gets or sets 休損型数</summary>
             /// <value>休損型数</value>
-            public int? LossAbsenceTypeCount { get; set; }
+            public decimal? LossAbsenceTypeCount { get; set; }
             /// <summary>Gets or sets 保全見解</summary>
             /// <value>保全見解</value>
             public string MaintenanceOpinion { get; set; }
